Added: Better file structure and separation of concerns
