+++ package.json
@@ -4,8 +4,8 @@
   "description": "Snake Server",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js",
+    "start": "node src/index.js",
+    "dev": "nodemon src/index.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "debug": "node --inspect index.js",
     "postinstall": "scripts/postinstall.sh"
