+++ src/objects/BodyPart.js
@@ -1,0 +1,10 @@
+const GameObject = require('./GameObject');
+
+class BodyPart extends GameObject {
+
+    constructor(game, position) {
+        super(game, position);
+    }
+}
+
+module.exports = BodyPart;
