+++ src/objects/Fruit.js
@@ -1,0 +1,10 @@
+const GameObject = require('./GameObject');
+
+class Fruit extends GameObject {
+
+    constructor(game, position) {
+        super(game, position);
+    }
+}
+
+module.exports = Fruit;
